Sometimes queue starts emptying after bot input
It seems as if tokens are not accumulating for some reason. Can't figure
out why. I don't think this will fix it, as Time.now < Time.now is true,
so our precision should be great enough to guarantee that once we wake
up from sleep, we need to add token + add new sleep.